// src/app/components/characters/character-description.enum.ts
export enum characterDescriptions {
    merlin = 'Knows all evil but Mordred',
    percival = 'Knows Merlin and Morgana',
    minionG = 'Does not know anything',
    mordred = 'Unknown to Merlin',
    morgana = 'Appears as Merlin',
    oberon = 'Unknown to Evil',
    assasin = 'Gets final decision',
    minionB = 'Knows other Evils',
    lancelot = 'Can be good or bad',
    lakeLady = 'I keep forgetting what this character does'
}